[Code Structure] Moved 5 years post to yaml data file.
