+++ source/blog/2014-12-21-5-years-of-ios.haml
@@ -6,19 +6,20 @@ timeline: true
 link_to: swift
 ---
 
-- require 'lib/5-years-of-ios.rb'
+- require 'lib/timeline.rb'
 
-%p In December, 2009, I started doing iOS development. That means that this month – almost to the day – I’ve been doing iOS for five years, meaning I’m finally qualified for an entry-level corporate job requiring 5-7 years experience. Kidding aside, I thought it would be nice to look back on the past half-decade of my life and pick out some events that were important to getting me where I am today. 
+%p In December, 2009, I started doing iOS development. That means that this month – almost to the day – I've been doing iOS for five years, meaning I'm finally qualified for an entry-level corporate job requiring 5-7 years experience. Kidding aside, I thought it would be nice to look back on the past half-decade of my life and pick out some events that were important to getting me where I am today.
+
 
 / <!-- more -->
 
-%p This isn’t meant to be a guide or advice in any way – this is just me reflecting on my experience and sharing it with you. I've had several personal revelations while writing this. It's been a really fun experience, but it’s a pretty personal post, so feel free to skip it.
+%p This isn't meant to be a guide or advice in any way – this is just me reflecting on my experience and sharing it with you. I've had several personal revelations while writing this. It's been a really fun experience, but it's a pretty personal post, so feel free to skip it.
 
-%p In the Fall of 2009, things seemed like they should be pretty good. In my third year of a five-year degree, I was a student volunteer: I helped with the student union to represent our faculty, the undergrad faculty association to organize social events, and as a peer mentor to assist first-year students with their homework. I was also doing some paid TA work to mark first-year assignments and help students in the computer labs. And finally, I was working part-time for Research in Motion. 
+%p In the Fall of 2009, things seemed like they should be pretty good. In my third year of a five-year degree, I was a student volunteer: I helped with the student union to represent our faculty, the undergrad faculty association to organize social events, and as a peer mentor to assist first-year students with their homework. I was also doing some paid TA work to mark first-year assignments and help students in the computer labs. And finally, I was working part-time for Research in Motion.
 
-%p But things were not good. My then fiancée had finally convinced me to see the university counsellor about some compulsive thoughts about hurting myself that I had been having for months. I had my intake appointment, but wouldn’t start counselling until after the Christmas break. 
+%p But things were not good. My then fiancée had finally convinced me to see the university counsellor about some compulsive thoughts about hurting myself that I had been having for months. I had my intake appointment, but wouldn't start counselling until after the Christmas break.
 
-%p Christmas breaks, I had learnt, are boring affairs. With no homework or term projects to work on, my mind would get restless. That year, we were spending the break with Ashley’s family in rural Nova Scotia. I wanted to spend some time learning to make iOS apps. Ashley agreed to let me borrow her MacBook (mine was a G4, so I couldn’t write iOS apps on it). 
+%p Christmas breaks, I had learnt, are boring affairs. With no homework or term projects to work on, my mind would get restless. That year, we were spending the break with Ashley's family in rural Nova Scotia. I wanted to spend some time learning to make iOS apps. Ashley agreed to let me borrow her MacBook (mine was a G4, so I couldn't write iOS apps on it).
 
 %p BEGIN_EXTRA_WIDE
 
@@ -26,8 +27,8 @@ link_to: swift
   %span.cntl-bar.cntl-center
     %span.cntl-bar-fill
   .cntl-states
-    = FiveYears::Events().map{ |e| "#{ e.to_html }\n" }.reduce(:+)
+    = Timeline::PrintEvents(data.five_years)
 
 %p END_EXTRA_WIDE
 
-Our Dutch visas expire in February and without a Dutch employer, we can’t stay any longer. Artsy is based in New York and Canadians have a pretty easy time getting work visas, so we’ll be moving there in February. As much as I will miss Amsterdam and all the friends I have here, I am looking forward to a new adventure. 
+%p Our Dutch visas expire in February and without a Dutch employer, we can't stay any longer. Artsy is based in New York and Canadians have a pretty easy time getting work visas, so we'll be moving there in February. As much as I will miss Amsterdam and all the friends I have here, I am looking forward to a new adventure.
